requirements.txt
pyjwt
Flask
Gunicorn
PYPDF2
Flask-Cors
pycryptodome